Remove changes related to MultiCouplingFourSolvers.
