--- tests/serial/multiCouplingFourSolvers/MultiCouplingFourSolvers2.cpp
@@ -1,22 +1,0 @@
-#ifndef PRECICE_NO_MPI
-
-#include <string>
-
-#include <fstream>
-#include "helpers.hpp"
-#include "testing/Testing.hpp"
-
-BOOST_AUTO_TEST_SUITE(PreciceTests)
-BOOST_AUTO_TEST_SUITE(Serial)
-BOOST_FIXTURE_TEST_SUITE(MultiCouplingFourSolvers, MultiCouplingFourSolversFixture)
-
-BOOST_AUTO_TEST_CASE(MultiCouplingFourSolvers2)
-{
-  PRECICE_TEST("SolverA"_on(1_rank), "SolverB"_on(1_rank), "SolverC"_on(1_rank), "SolverD"_on(1_rank));
-  multiCouplingFourSolvers(context.config(), context);
-}
-
-BOOST_AUTO_TEST_SUITE_END()
-BOOST_AUTO_TEST_SUITE_END()
-BOOST_AUTO_TEST_SUITE_END()
-#endif // PRECICE_NO_MPI
