+++ tests/tests.cmake
@@ -3,8 +3,4 @@
 #
 target_sources(testprecice
     PRIVATE
-    tests/serial/multiCouplingFourSolvers/MultiCouplingFourSolvers1.cpp
-    tests/serial/multiCouplingFourSolvers/MultiCouplingFourSolvers2.cpp
-    tests/serial/multiCouplingFourSolvers/helpers.cpp
-    tests/serial/multiCouplingFourSolvers/helpers.hpp
     )
